Split formatter method into multiple files.
