+++ Cargo.toml
@@ -11,6 +11,7 @@ version = "0.2.1"
 ignore = "0.4"
 pest = "2.5"
 pest_derive = "2.5"
+pest_meta = "2.5"
 text-utils = "0.2"
 toml = "0.5"
